Readme: Small update, friendly link to unifi-api
Currently the code is here, but unifi-api seems to provide a cleaner
approach. For the sake of history the code stays in place
(if someone wants to contribute, go for it).

It has been last tested against 2.3.9, no idea if it will still work
with the upcoming 3.0 release, UBNT hasn't yet published changes to the API.